+++ bikeability/indicators/__init__.py
@@ -1,0 +1,3 @@
+from .path_categories import *
+from .smoothness import *
+from .surface_types import *