+++ test/test_operator_worker.py
@@ -1,30 +1,9 @@
-import geopandas as gpd
-import shapely
 from geopandas import testing
 
-from bikeability.utils import (
-    PathCategory,
-)
 
-
-def test_get_paths(operator, expected_compute_input, ohsome_api):
-    line_geom = shapely.LineString([(12.3, 48.22), (12.3, 48.2205), (12.3005, 48.22)])
-    polygon_geom = shapely.Polygon(((12.3, 48.22), (12.3, 48.2205), (12.3005, 48.22), (12.3, 48.22)))
-
-    expected_lines = gpd.GeoDataFrame(
-        data={
-            'geometry': [line_geom],
-            '@other_tags': [{'bicycle': 'no'}],
-        },
-        crs='EPSG:4326',
-    )
-    expected_polygons = gpd.GeoDataFrame(
-        data={
-            'geometry': [polygon_geom],
-            '@other_tags': [{'bicycle': 'no'}],
-        },
-        crs='EPSG:4326',
-    )
+def test_get_paths(operator, expected_compute_input, ohsome_api, test_line, test_polygon):
+    expected_lines = test_line.drop(['category', 'rating'], axis=1)
+    expected_polygons = test_polygon.drop(['category', 'rating'], axis=1)
     computed_lines, computed_polygons = operator.get_paths(expected_compute_input.get_aoi_geom())
 
     testing.assert_geodataframe_equal(
@@ -41,9 +20,3 @@ def test_get_paths(operator, expected_compute_input, ohsome_api):
         check_geom_type=True,
         check_less_precise=True,
     )
-
-
-def test_input_categories_match_PathCategories(expected_compute_input):
-    input_categories = expected_compute_input.get_path_rating_mapping()
-
-    assert set(input_categories.keys()) == set(PathCategory)